+++ .docgenirc.js
@@ -3,8 +3,12 @@ module.exports = {
     title: 'NgxGantt',
     siteProjectName: 'example',
     outputDir: 'dist/site',
-    logoUrl: 'https://cdn.pingcode.com/open-sources/gantt/logo.png',
+    logoUrl: '/assets/imgs/logo.png',
     repoUrl: 'https://github.com/worktile/ngx-gantt',
+    locales: [
+        { key: 'en-us', name: 'English' },
+        { key: 'zh-cn', name: '中文' }
+    ],
     defaultLocale: 'zh-cn',
     navs: [
         null,