+++ packages/gantt/src/components/baseline/baseline.component.ts
@@ -1,15 +1,20 @@
-import { Component, ElementRef, HostBinding, Inject, Input, OnInit, ViewChild } from '@angular/core';
+import { Component, ElementRef, HostBinding, Inject, Input, OnInit, TemplateRef, ViewChild } from '@angular/core';
 import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 import { GanttBaselineItemInternal } from '../../class/baseline';
 import { GanttUpper, GANTT_UPPER_TOKEN } from '../../gantt-upper';
+import { NgIf, NgTemplateOutlet } from '@angular/common';
 
 @Component({
     selector: 'ngx-gantt-baseline,gantt-baseline',
-    templateUrl: './baseline.component.html'
+    templateUrl: './baseline.component.html',
+    standalone: true,
+    imports: [NgIf, NgTemplateOutlet]
 })
 export class NgxGanttBaselineComponent implements OnInit {
     @Input() baselineItem: GanttBaselineItemInternal;
+
+    @Input() template: TemplateRef<any>;
 
     public unsubscribe$ = new Subject<void>();